lava_car/requirements.txt
fontawesome-free==5.15.4
django-bootstrap-v5==1.0.8
django-bootstrap-form==3.4